Testing: Add test for set_exceptions_enabled
test/t_java.java: Tests exception handling. This tests a test case known
to fail (t_bounded.cxx on LD_SLSQP) with set_exceptions_enabled set to
true if the argument is 1, false if the argument is 0, and verifies that
the exception is thrown resp. not thrown, as expected.

test/CMakeLists.txt: Add the C++ test t_except with the instances 1
(true = exceptions enabled) and 0 (false = exceptions disabled). Both
instances pass.